Make rainbow coloured trees!
So I've now figured out *a way* of setting colours on cubes. For now, I am selecting the colour of the cube based on it's `treeIndex`, so each tree/shrub is a different colour from the choices from the LGBT Pride flag.

They are rotating based on the SawLFO, but this is obviously not the correct way to do this, that's my next job.